refactor(rust): extract gridfs + version into db/ (slice 4/9)
Move GridFS browsing (GridFsFileInfo, list/download_gridfs_file_impl)
into db/gridfs.rs and get_mongodb_version_impl into db/version.rs;
lib.rs re-exports them so command wrappers and tests are unchanged.
cargo test green (60), no warnings.
