+++ src-tauri/src/db/mod.rs
@@ -1,3 +1,5 @@
 //! Database-operation modules (split out of lib.rs).
 
+pub mod gridfs;
 pub mod schema;
+pub mod version;
